refactor: required data와 not required data를 나누기 위해 API 별로 DTO를 분리"
