+++ backend/src/main/java/study/memoravel/config/MailConfig.java
@@ -1,4 +1,4 @@
-package study.memoravel;
+package study.memoravel.config;
 
 
 import org.springframework.beans.factory.annotation.Value;
