+++ backend/src/main/java/study/memoravel/config/SpringConfig.java
@@ -1,4 +1,4 @@
-package study.memoravel;
+package study.memoravel.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
