+++ backend/src/main/java/study/memoravel/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package study.memoravel;
+package study.memoravel.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
