+++ backend/src/main/java/study/memoravel/controller/account/LoginInfo.java
@@ -1,0 +1,20 @@
+package study.memoravel.controller.account;
+
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@Builder
+@ApiModel(value = "계정 로그인 요청")
+public class LoginInfo {
+    @ApiModelProperty(value = "이메일")
+    private String email;
+    @ApiModelProperty(value = "비밀번호")
+    private String password;
+}
