+++ backend/src/main/java/study/memoravel/controller/account/SignupInfo.java
@@ -1,0 +1,27 @@
+package study.memoravel.controller.account;
+
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+
+@Getter
+@Setter
+@AllArgsConstructor
+@Builder
+@ApiModel("계정 가입 정보")
+public class SignupInfo {
+    @ApiModelProperty(value = "이메일")
+    private String email;
+    @ApiModelProperty(value = "별명")
+    private String nickname;
+    @ApiModelProperty(value = "비밀번호")
+    private String password;
+    @ApiModelProperty(value = "주소")
+    private String address;
+    @ApiModelProperty(value = "성별")
+    private String gender;
+}
