+++ backend/src/main/java/study/memoravel/service/UserService.java
@@ -2,8 +2,9 @@ package study.memoravel.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-import study.memoravel.domain.Login;
-import study.memoravel.domain.User;
+import study.memoravel.controller.account.LoginInfo;
+import study.memoravel.controller.account.SignupInfo;
+import study.memoravel.controller.userInfo.UserInfo;
 import study.memoravel.repository.UserRepository;
 import study.memoravel.util.Encoding;
 import study.memoravel.util.JWT;
@@ -17,34 +18,34 @@ public class UserService {
         UserService.userRepo = userRepo;
     }
 
-    public String login(Login.DTO loginInfo) throws Exception {
-        User.DTO result = userRepo.findByEmail(loginInfo.getEmail());
+    public String login(LoginInfo login) throws Exception {
+        UserInfo result = userRepo.findByEmail(login.getEmail());
         if (result == null) {
             throw new Exception("Failed Login");
         }
-        if (Encoding.checkBCrypt(loginInfo.getPassword(), result.getPassword())) {
-            return JWT.create(result);
+        if (Encoding.checkBCrypt(login.getPassword(), result.getPassword())) {
+            return JWT.create(result.getNickname(), result.getEmail());
         } else {
             throw new Exception("Failed Login");
         }
     }
 
-    public String signup(User.DTO user) {
+    public String signup(SignupInfo user) {
         user.setPassword(Encoding.getBCrypt(user.getPassword()));
         userRepo.save(user);
-        return JWT.create(user);
+        return JWT.create(user.getNickname(), user.getEmail());
     }
 
-    public User.DTO getUser(String email) {
+    public UserInfo getUser(String email) {
         return userRepo.findByEmail(email);
     }
 
-    public void setUser(String email, User.DTO user) {
-        userRepo.updateUser(email, user);
+    public void setUser(String email, UserInfo userInfo) {
+        userRepo.updateUser(email, userInfo);
     }
 
     public Boolean checkPhoneNumber(String phoneNumber) {
-        User.DTO result = userRepo.findByPhoneNumber(phoneNumber);
+        UserInfo result = userRepo.findByPhoneNumber(phoneNumber);
         return result == null;
     }
 
@@ -53,7 +54,7 @@ public class UserService {
     }
 
     public Boolean checkEmail(String mail) {
-        User.DTO result = userRepo.findByEmail(mail);
+        UserInfo result = userRepo.findByEmail(mail);
         return result == null;
     }
 }
