+++ backend/src/main/java/study/memoravel/util/JWT.java
@@ -1,7 +1,6 @@
 package study.memoravel.util;
 
 import io.jsonwebtoken.*;
-import study.memoravel.domain.User;
 
 import java.io.IOException;
 import java.util.Date;
@@ -22,7 +21,7 @@ public class JWT {
         return result;
     }
 
-    public static String create(User.DTO userInfo) {
+    public static String create(String nickname, String email) {
         if (secret == null) {
             secret = getSecret();
         }
@@ -30,11 +29,11 @@ public class JWT {
         long expiredTime = 1000 * 60L * 30L; // 30분
 
         return Jwts.builder().setHeaderParam(Header.TYPE, Header.JWT_TYPE) // header 설정
-                .setIssuer(userInfo.getNickname()) // 발급자 설정
+                .setIssuer(nickname) // 발급자 설정
                 .setIssuedAt(now) // 발급 시간 설정
                 .setExpiration(new Date(now.getTime() + expiredTime)) // 만료 시간 설정
-                .claim("email", userInfo.getEmail()) // 비공개 클레임 설정(ID만 사용)
-                .claim("trash", new Date().getTime())
+                .claim("email", email) // 비공개 클레임 설정(ID만 사용)
+                .claim("date", new Date().getTime())
                 .signWith(SignatureAlgorithm.HS256, secret) // 해싱 알고리즘과 시크릿 키 설정
                 .compact();
     }
